fix(codeql): resolve 6 security/correctness alerts (M1)
- dns_wrap.cpp: fix suspicious-allocation-size (malloc sizeof)
- p2p_proxy_server: fix stack-address-escape (unique_ptr)
- ryu_protocol.hpp: suppress stack-address-escape (zero-copy design)
- main.cpp: fix integer-multiplication-cast-to-long
- config.cpp, log.cpp: suppress world-writable (FAT32 no POSIX perms)

+++ sysmodule/source/main.cpp
@@ -48,7 +48,7 @@ namespace ams {
             const u32 sum = tcp_tx_buf_max_size + tcp_rx_buf_max_size +
                             config->udp_tx_buf_size + config->udp_rx_buf_size;
 
-            return config->sb_efficiency * util::AlignUp(sum, os::MemoryPageSize);
+            return static_cast<size_t>(config->sb_efficiency) * util::AlignUp(sum, os::MemoryPageSize);
         }
 
         /// Socket initialization configuration
